# src/main/resources/finance_openapi.yaml
openapi: 3.0.3
info:
  title: Personal Finance Tracker API
  version: 1.0.0
  description: API для приложения учета личных финансов
servers:
  - url: http://localhost:8080
    description: Локальный сервер
  - url: http://103.76.52.13:8080
    description: Удаленный тестовый сервер
paths:
...
  /api/v1/transactions:
    post:
      summary: Добавить новую транзакцию с поддержкой идемпотентности
      security:
        - BearerAuth: []
      parameters:
        - name: X-Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Транзакция успешно добавлена (или возвращена из кэша идемпотентности)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionInfoDto'
        '400':
          description: Ошибка бизнес-логики или валидации при создании транзакции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                InsufficientFunds:
                  summary: Нехватка средств (овердрафт запрещен)
                  value:
                    error: "Insufficient Funds"
                    message: "Недостаточно средств на счете. Баланс: 0.00, Попытка списания: 5000.0"
                    status: 400
                    timestamp: "2026-06-20 12:00:00"
                CurrencyMismatch:
                  summary: Несовпадение валюты операции и счёта
                  value:
                    error: "Currency Mismatch"
                    message: "Несовпадение валюты. Валюта счёта ID = 1 - RUB, валюта операции - USD"
                    status: 400
                    timestamp: "2026-05-29 12:00:00"
                ValidationFailed:
                  summary: Ошибка формата данных или отсутствие заголовка
                  value:
                    error: "Validation Failed"
                    message: "Idempotency key is missing / amount must be greater than 0"
                    status: 400
                    timestamp: "2026-05-29 12:00:00"
        '404':
          description: Указанный в теле запроса accountId не найден, либо этот счёт не принадлежит текущему пользователю.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error: "Resource Not Found"
                message: "Счет не найден с id: 1"
                status: 404
                timestamp: "2026-05-29 12:00:00"
        '409':
          description: |
            Конфликт идемпотентности (Duplicate request). 
            Если для данного ключа в Redis уже существует запись ответа idempotency-key (хранится 12 часов),
            API автоматически вернет сохраненный успешный ответ (статус 201).
            
            Если ответа в кэше нет,
            система пытается установить распределенную блокировку idempotency-key на 180 секунд.
            
            Статус 409 возвращается в случае, если распределенная блокировка уже удерживается другим параллельным потоком.
            Это означает, что идентичный запрос прямо сейчас находится в процессе обработки базой данных.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error: "Idempotency key already in use"
                message: "Duplicate request in progress"
                status: 409
                timestamp: "2026-06-24T17:15:00.000Z"
  /api/v1/transactions/{transactionId}:
    put:
      summary: Изменить существующую транзакцию
      security:
        - BearerAuth: [ ]
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdateRequest'
      responses:
        '200':
          description: Транзакция успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionInfoDto'
    delete:
      summary: Удалить транзакцию
      security:
        - BearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Транзакция удалена
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TransactionRequest:
      type: object
      required: [ accountId, type, amount, currency, categoryId, date ]
      properties:
        accountId:
          type: integer
          format: int64
          minimum: 1
          example: 1
        type:
          type: string
          enum: [ INCOME, EXPENSE ]
          example: "EXPENSE"
        amount:
          type: number
          format: decimal
          minimum: 0.01
          example: 1500.00
        currency:
          type: string
          enum: [ RUB, EUR, USD ]
          example: "RUB"
        categoryId:
          type: integer
          format: int64
          minimum: 1
          example: 2
        date:
          type: string
          format: date-time
          example: "2026-06-24T14:30:00.000Z"
        description:
          type: string
          maxLength: 300
          example: "Покупка продуктов"
    TransactionUpdateRequest:
      type: object
      properties:
        amount:
          type: number
          format: decimal
          minimum: 0.01
          example: 1800.00
        type:
          type: string
          enum: [ INCOME, EXPENSE ]
        categoryId:
          type: integer
          format: int64
          minimum: 1
        date:
          type: string
          format: date-time
          example: "2026-06-24T15:00:00.000Z"
        description:
          type: string
          maxLength: 300
    TransactionInfoDto:
      type: object
      properties:
        transactionId:
          type: integer
          format: int64
          example: 105
        accountId:
          type: integer
          format: int64
          example: 1
        type:
          type: string
          enum: [ INCOME, EXPENSE ]
        amount:
          type: number
          format: decimal
          example: 1500.00
        category:
          type: string
          example: "Продукты"
        date:
          type: string
          format: date-time
          example: "2026-06-24T14:30:00.000Z"
        description:
          type: string
          example: "Покупка продуктов"
        accountBalance:
          type: number
          format: decimal
          example: 98500.00
...
    ApiError:
      type: object
      required: [ error, message, status, timestamp ]
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time